chore: reduce registry code bloat and improve DI (#3794)
